requirements.txt
# curateur - ScreenScraper ROM Metadata & Media Scraper
# Python package dependencies

# Core dependencies
PyYAML>=6.0
lxml>=4.9.0
requests>=2.31.0
Pillow>=10.0.0
rich>=13.0.0
psutil>=5.9.0

# Development dependencies (optional)
# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
responses>=0.23.0
